[Silver III] Title: 주유소, Time: 80 ms, Memory: 3588 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -4,11 +4,11 @@
 
 using namespace std;
 
-int m;
+long long m;
 long long ans;
 int n;
-vector<int> dis;
-vector<int>node;
+vector<long long> dis;
+vector<long long>node;
 int main() {
 	cin >> n;
 	dis.resize(n - 1);
